requirements.txt
Flask
Flask-SQLAlchemy
Flask-RESTful
Flask-Marshmallow
marshmallow-sqlalchemy
openai
python-dotenv
Werkzeug
Pillow